+++ example/ios/Runner/AppDelegate.swift
@@ -7,7 +7,7 @@ import flutter_callkit_incoming
 
 @UIApplicationMain
 @objc class AppDelegate: FlutterAppDelegate, PKPushRegistryDelegate, CallkitIncomingAppDelegate {
-
+    
     
     override func application(
         _ application: UIApplication,
@@ -20,13 +20,40 @@ import flutter_callkit_incoming
         let voipRegistry: PKPushRegistry = PKPushRegistry(queue: mainQueue)
         voipRegistry.delegate = self
         voipRegistry.desiredPushTypes = [PKPushType.voIP]
-
+        
         //Use if using WebRTC
         //RTCAudioSession.sharedInstance().useManualAudio = true
         //RTCAudioSession.sharedInstance().isAudioEnabled = false
         
+        //Add for Missed call notification
+        if #available(iOS 10.0, *) {
+            UNUserNotificationCenter.current().delegate = self as UNUserNotificationCenterDelegate
+        }
+        
         return super.application(application, didFinishLaunchingWithOptions: launchOptions)
     }
+    
+    // Add for Missed call notification(show notification when foreground)
+    override func userNotificationCenter(_ center: UNUserNotificationCenter,
+                                         willPresent notification: UNNotification,
+                                         withCompletionHandler completionHandler:
+                                         @escaping (UNNotificationPresentationOptions) -> Void) {
+        
+        CallkitNotificationManager.shared.userNotificationCenter(center, willPresent: notification, withCompletionHandler: completionHandler)
+    }
+    
+    // Add for Missed call notification(action when click callback in missed notification)
+    override func userNotificationCenter(_ center: UNUserNotificationCenter,
+                                         didReceive response: UNNotificationResponse,
+                                         withCompletionHandler completionHandler: @escaping () -> Void) {
+        if response.actionIdentifier == CallkitNotificationManager.CALLBACK_ACTION {
+            let data = response.notification.request.content.userInfo as? [String: Any]
+            SwiftFlutterCallkitIncomingPlugin.sharedInstance?.sendCallbackEvent(data)
+        }
+        completionHandler()
+    }
+    
+    
     
     // Call back from Recent history
     override func application(_ application: UIApplication,
@@ -97,7 +124,7 @@ import flutter_callkit_incoming
                 print("Received data: \(data)")
                 //Make sure call action.fulfill() when you are done(connected WebRTC - Start counting seconds)
                 action.fulfill()
-
+                
             case .failure(let error):
                 print("Error: \(error.localizedDescription)")
             }
@@ -114,7 +141,7 @@ import flutter_callkit_incoming
                 print("Received data: \(data)")
                 //Make sure call action.fulfill() when you are done
                 action.fulfill()
-
+                
             case .failure(let error):
                 print("Error: \(error.localizedDescription)")
             }
@@ -131,7 +158,7 @@ import flutter_callkit_incoming
                 print("Received data: \(data)")
                 //Make sure call action.fulfill() when you are done
                 action.fulfill()
-
+                
             case .failure(let error):
                 print("Error: \(error.localizedDescription)")
             }
@@ -146,7 +173,7 @@ import flutter_callkit_incoming
             switch result {
             case .success(let data):
                 print("Received data: \(data)")
-
+                
             case .failure(let error):
                 print("Error: \(error.localizedDescription)")
             }
@@ -192,7 +219,7 @@ import flutter_callkit_incoming
                     completion(.failure(NSError(domain: "mobile.app", code: 0, userInfo: [NSLocalizedDescriptionKey: "Empty data"])))
                     return
                 }
-
+                
                 do {
                     let jsonObject = try JSONSerialization.jsonObject(with: data, options: [])
                     completion(.success(jsonObject))