+++ ios/Classes/Call.swift
@@ -160,6 +160,10 @@ public class Call: NSObject {
     @objc public var isShowMissedCallNotification: Bool = true
     @objc public var missedNotificationSubtitle: String
     
+    @objc public var missedNotificationCallbackText: String
+    @objc public var isShowCallback: Bool = true
+    
+    
     @objc public init(id: String, nameCaller: String, handle: String, type: Int) {
         self.uuid = id
         self.nameCaller = nameCaller
@@ -190,6 +194,8 @@ public class Call: NSObject {
         
         self.isShowMissedCallNotification = true
         self.missedNotificationSubtitle = "Missed Call"
+        self.missedNotificationCallbackText = "Call back"
+        self.isShowCallback = true
     }
     
     @objc public convenience init(args: NSDictionary) {
@@ -249,18 +255,24 @@ public class Call: NSObject {
             self.audioSessionPreferredIOBufferDuration = args["audioSessionPreferredIOBufferDuration"] as? Double ?? 0.005
         }
         if let missedCallNotification = args["missedCallNotification"] as? [String: Any] {
-            self.isShowMissedCallNotification = missedCallNotification["isShowMissedCallNotification"] as? Bool ?? true
-            self.missedNotificationSubtitle = missedCallNotification["missedNotificationSubtitle"] as? String ?? "Missed Call"
+            self.isShowMissedCallNotification = missedCallNotification["showNotification"] as? Bool ?? true
+            self.missedNotificationSubtitle = missedCallNotification["subtitle"] as? String ?? "Missed Call"
+            self.missedNotificationCallbackText = missedCallNotification["callbackText"] as? String ?? "Call back"
+            self.isShowCallback = missedCallNotification["isShowCallback"] as? Bool ?? true
         }else {
             self.isShowMissedCallNotification = true
             self.missedNotificationSubtitle = "Missed Call"
+            self.missedNotificationCallbackText = "Call back"
+            self.isShowCallback = true
         }
     }
     
     open func toJSON() -> [String: Any] {
         let missedCallNotification: [String : Any] = [
             "showNotification": isShowMissedCallNotification,
-            "missedNotificationSubtitle": missedNotificationSubtitle
+            "subtitle": missedNotificationSubtitle,
+            "callbackText": missedNotificationCallbackText,
+            "isShowCallback": isShowCallback
         ]
         let ios: [String : Any] = [
             "iconName": iconName,